2 commits in one "next pending event" - "Prevent missing servicing"
Needed "msm: kgsl: Make the GPU device aware of the next pending event" to make "msm: kgsl: Prevent missing servicing of events waiting on timestamps" Work

msm: kgsl: Make the GPU device aware of the next pending event
The adreno core needs to know what the next event pending for
any given context is so it can mark the interupt to be fired.
If this isn't done then some timestamps that don't have a
matching waittimestamp call won't fire an interrupt. This is
dangerous on the last interrupt/event before a context goes
away.

Change-Id: Ic0dedbad71f6de07b43b0656128c76509326d645
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>

----------------------------------------------------------------------------------------------------

msm: kgsl: Prevent missing servicing of events waiting on timestamps
Prevent instances where we miss servicing an event which was waiting
on a timestamp. We do this by looping in the events list, until
we finish servicing all the members of the list, and if we come across
an event which has not been serviced yet, we write a dummy interrupt to
the ringbuffer so that we don't miss when that timestamp's work is done.

Change-Id: Ib7b3554efa0d9a8bdc20d1722d912977f72f3ed5
Signed-off-by: Harsh Vardhan Dwivedi <hdwivedi@codeaurora.org>

Conflicts:

	drivers/gpu/msm/adreno.c
	drivers/gpu/msm/kgsl.c
	drivers/gpu/msm/kgsl_device.h